feat(as-3619): removed hardcoded data, made page dynamic and link pages
BREAKING CHANGE: URLs of validation officer journey has been changed
