+++ packages/web-app/src/controllers/questionnaire-check-and-confirm.js
@@ -1,10 +1,15 @@
+const {
+  reviewQuestionnaireSubmission: previousPage,
+  reviewQuestionnaireComplete: nextPage,
+} = require('../config/views');
 const getCheckAndConfirm = (req, res) => {
   const { questionnaire } = req.session;
 
   res.render('questionnaire-check-and-confirm', {
     pageTitle: 'Review questionnaire',
+    previousPage: `${previousPage}/${questionnaire.appealId}`,
     questionnaireData: questionnaire,
-    reviewOutcome: 'Incomplete',
+    reviewOutcome: questionnaire.outcome,
   });
 };
 
@@ -14,4 +19,5 @@ const postCheckAndConfirm = (req, res) => {
 
 module.exports = {
   getCheckAndConfirm,
+  postCheckAndConfirm,
 };
