+++ packages/web-app/src/routes/review-questionnaire-submission.test.js
@@ -1,25 +1,27 @@
 const {
-  getReviewQuestionnaire,
-  postReviewQuestionnaire,
-} = require('../controllers/review-questionnaire');
+  getReviewQuestionnaireSubmission,
+  postReviewQuestionnaireSubmission,
+} = require('../controllers/review-questionnaire-submission');
 
-const { rules: reviewQuestionnaireValidationRules } = require('../validation/review-questionnaire');
+const {
+  rules: reviewQuestionnaireValidationRules,
+} = require('../validation/review-questionnaire-submission');
 const expressValidationErrorsToGovUkErrorList = require('../lib/express-validation-errors-to-govuk-error-list');
 
-jest.mock('../validation/review-questionnaire');
+jest.mock('../validation/review-questionnaire-submission');
 
 const { mockGet, mockPost } = require('../../test/utils/mocks');
 
 describe('routes/questionnaires-for-review/review', () => {
   it('should define the correct route', () => {
     // eslint-disable-next-line global-require
-    require('./review-questionnaire');
-    expect(mockGet).toBeCalledWith(`/`, getReviewQuestionnaire);
+    require('./review-questionnaire-submission');
+    expect(mockGet).toBeCalledWith(`/`, getReviewQuestionnaireSubmission);
     expect(mockPost).toBeCalledWith(
       `/`,
       reviewQuestionnaireValidationRules(),
       expressValidationErrorsToGovUkErrorList,
-      postReviewQuestionnaire
+      postReviewQuestionnaireSubmission
     );
   });
 });
