+++ packages/web-app/src/views/components/missing-or-incorrect.njk
@@ -1,0 +1,9 @@
+{% macro missingOrIncorrect(items) %}
+  <ul class="govuk-list">
+
+    {% for item in items %}
+      <li>{{ item }}</li>
+    {% endfor %}
+
+  </ul>
+{% endmacro %}
