+++ packages/web-app/src/views/review-questionnaire-complete.njk
@@ -7,7 +7,7 @@
     <div class="govuk-panel govuk-panel--confirmation govuk-!-margin-bottom-7">
         {{ govukPanel({
             titleText: "Outcome confirmed",
-            html: "Questionnaire " + status + "<br><strong>" + appealData.caseReference + "</strong>"
+            html: "Questionnaire " + outcome + "<br><strong>" + appealData.caseReference + "</strong>"
         })  }}
     </div>
